+++ Server/KFContrib/KFNetwork/KFNetConnector.h
@@ -14,14 +14,14 @@ namespace KFrame
         ~KFNetConnector();
 
         // 初始化
-        void InitConnector( uint32 id, KFNetServices* netservices );
+        void InitConnector( uint64 id, KFNetServices* netservices );
 
         // 弹出一个收到消息
         KFNetMessage* PopNetMessage();
 
         // 发送消息
         bool SendNetMessage( uint32 msgid, const char* data, uint32 length );
-        bool SendNetMessage( uint32 objectid, uint32 msgid, const char* data, uint32 length );
+        bool SendNetMessage( uint64 objectid, uint32 msgid, const char* data, uint32 length );
 
         // 执行消息处理
         void RunUpdate( KFNetFunction& netfunction, uint32 maxcount );
@@ -31,10 +31,10 @@ namespace KFrame
         KFNetMessage* PopMessage();
 
         // 发送单一消息
-        bool SendSingleMessage( uint32 objectid, uint32 msgid, const char* data, uint32 length );
+        bool SendSingleMessage( uint64 objectid, uint32 msgid, const char* data, uint32 length );
 
         // 发送拆包消息
-        bool SendMultiMessage( uint32 objectid, uint32 msgid, const char* data, uint32 length );
+        bool SendMultiMessage( uint64 objectid, uint32 msgid, const char* data, uint32 length );
         //////////////////////////////////////////////////////////////////////////////////////////////////
         //////////////////////////////////////////////////////////////////////////////////////////////////
         // 发送ping逻辑
