+++ Server/KFPlugin/KFClusterClient/KFClusterClientModule.h
@@ -34,14 +34,14 @@ namespace KFrame
         ////////////////////////////////////////////////////////////////////////////////
         ////////////////////////////////////////////////////////////////////////////////
         // 调用注册函数
-        void CallClusterConnectionFunction( const std::string& name, uint32 serverid );
+        void CallClusterConnectionFunction( const std::string& name, uint64 serverid );
 
         // 发送消息
         virtual bool SendToShard( const std::string& name, uint32 msgid, google::protobuf::Message* message );
-        virtual bool SendToShard( const std::string& name, uint32 shardid, uint32 msgid, google::protobuf::Message* message );
+        virtual bool SendToShard( const std::string& name, uint64 shardid, uint32 msgid, google::protobuf::Message* message );
 
         // 发送到静态对象所在的分片服务器
-        virtual bool SendToStaticObject( const std::string& name, uint32 objectid, uint32 msgid, google::protobuf::Message* message );
+        virtual bool SendToStaticObject( const std::string& name, uint64 objectid, uint32 msgid, google::protobuf::Message* message );
 
         // 发送到动态对象所在的分片服务器
         virtual bool SendToDynamicObject( const std::string& name, uint64 objectid, uint32 msgid, google::protobuf::Message* message );
@@ -67,7 +67,7 @@ namespace KFrame
 
         // 发送消息
         bool SendNetMessage( const std::string& name, uint32 msgid, google::protobuf::Message* message );
-        bool SendNetMessage( const std::string& name, uint32 shardid, uint32 msgid, google::protobuf::Message* message );
+        bool SendNetMessage( const std::string& name, uint64 shardid, uint32 msgid, google::protobuf::Message* message );
     private:
         // 集群客户端列表
         KFMap< std::string, const std::string&, KFClusterClient > _kf_cluster_client;
